+++ openapi/views.py
@@ -5,7 +5,7 @@ from rest_framework.views import APIView
 from data.models import GuName as GuNameModel
 from .serializers import OpenAPISewerPipeModelSerializer
 
-from .tests import save_sewerpipe_data_all_gubn, save_sewerpipe_data, save_rainfall_data
+from .utils import save_sewerpipe_data_all_gubn, save_sewerpipe_data, save_rainfall_data
 
 
 # url : openapi/data/save-previous-sewerpipe-data/<start_date>/<end_date>/
